+++ src/main/groovy/com/moniday/dto/AccountDTO.groovy
@@ -3,7 +3,7 @@ package com.moniday.dto
 class AccountDTO {
     String typeOfAccount
     String accountNumber
-    Long balance
+    BigDecimal balance
     String currencyType
     String deductedMoney = ""
 
@@ -16,7 +16,7 @@ class AccountDTO {
     AccountDTO(Map accountMap) {
         this.typeOfAccount = accountMap.typeOfAccount
         this.accountNumber = accountMap.accountNumber
-        this.balance = (accountMap.balance as Long)
+        this.balance = (accountMap.balance as BigDecimal)
         this.currencyType = accountMap.currencyType
 
         accountMap.transactions.each {
